lintcheck: move out of clippy-dev into own crate
