+++ clippy_dev/src/lib.rs
@@ -12,7 +12,6 @@ use walkdir::WalkDir;
 
 pub mod bless;
 pub mod fmt;
-pub mod lintcheck;
 pub mod new_lint;
 pub mod ra_setup;
 pub mod serve;
