+++ lintcheck/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "lintcheck"
+version = "0.0.1"
+authors = ["Matthias Krüger <matthias.krueger@famsik.de>"]
+edition = "2018"
+
+[dependencies]
+clap = "2.33"
+flate2 = {version = "1.0.19"}
+fs_extra = {version = "1.2.0"}
+rayon = {version = "1.5.0"}
+serde = {version = "1.0", features = ["derive"]}
+serde_json = {version = "1.0"}
+tar = {version = "0.4.30"}
+toml = {version = "0.5"}
+ureq = {version = "2.0.0-rc3"}
+
+[features]
+deny-warnings = []
